cambios x2

+++ Fianzas_Personales/app/src/main/java/campa/aracely/fianzas_personales/utilities/Transaccion.kt
@@ -3,7 +3,7 @@ package campa.aracely.fianzas_personales.utilities
 data class Transaccion(
     var id: String = "",
     val categoria: String = "",
-    val cantidad: Double = 0.0,
+    val cantidad: Float = 0.0f,
     val descripcion: String = "",
     val fecha: String = "",
     val tipoGasto: String = ""
